refactor: move table topbar into page component
